+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/MobArenaManager.java
@@ -5,12 +5,19 @@ import com.garbagemule.MobArena.util.ItemParser;
 import com.garbagemule.MobArena.util.ItemProvider;
 import org.bukkit.inventory.ItemStack;
 
+import java.util.Set;
+
 public class MobArenaManager implements ItemProvider {
     private final MageController controller;
 
     public MobArenaManager(MageController controller) {
         this.controller = controller;
         ItemParser.registerItemProvider(this);
+
+        Set<String> magicMobKeys = controller.getMobKeys();
+        for (String mob : magicMobKeys) {
+            new MagicMACreature(controller, mob, controller.getMob(mob));
+        }
     }
 
     @Override